feat: Change tolerance type to Double for numerical questions and update related validations

+++ backend/src/main/java/cz/scrumdojo/quizmaster/question/Question.java
@@ -42,5 +42,5 @@ public class Question {
     private String imageUrl;
 
     @Column(name = "tolerance")
-    private Integer tolerance;
+    private Double tolerance;
 }
